modified uniprot data download pipeline
